Unit Tests For CompanyServiceImpl and Feign Client for Authorization

+++ E-StockMarketApplication/src/main/java/com/example/EStockMarketApplication/Controllers/CompanyController.java
@@ -2,6 +2,7 @@ package com.example.EStockMarketApplication.Controllers;
 
 import com.example.EStockMarketApplication.DTOs.CompanyResponseDTO;
 import com.example.EStockMarketApplication.Exceptions.CompanyNotFound;
+import com.example.EStockMarketApplication.Feign.AuthorizeClient;
 import com.example.EStockMarketApplication.Models.Company;
 import com.example.EStockMarketApplication.Service.CompanyService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,37 +17,65 @@ import java.util.Optional;
 @RequestMapping("/v1.0/market/company")
 public class CompanyController {
     private CompanyService companyService;
+    private AuthorizeClient authorizeClient;
 
     @Autowired
-    public CompanyController(CompanyService companyService) {
+    public CompanyController(CompanyService companyService,AuthorizeClient authorizeClient) {
         this.companyService = companyService;
+        this.authorizeClient=authorizeClient;
     }
 
     @GetMapping("/info/{companyCode}")
-    public ResponseEntity<Optional<CompanyResponseDTO>> getCompanyInfo(@PathVariable Long companyCode) {
-        Optional<CompanyResponseDTO> company = companyService.getCompanyByID(companyCode);
-        return ResponseEntity.ok(company);
+    public ResponseEntity<?> getCompanyInfo(@PathVariable Long companyCode,@RequestHeader ("Authorization") String token) {
+        if(authorizeClient.authorize(token)) {
+            Optional<CompanyResponseDTO> company = companyService.getCompanyByID(companyCode);
+            return ResponseEntity.ok(company);
+        }
+        else
+        {
+            return new ResponseEntity<>("User not Authenticated",HttpStatus.BAD_REQUEST);
+        }
+
     }
 
     @GetMapping("/getAll")
-    public ResponseEntity<List<CompanyResponseDTO>> getAllCompanies() {
-        List<CompanyResponseDTO> companies = companyService.getAllCompanies();
-        return ResponseEntity.ok(companies);
+    public ResponseEntity<?> getAllCompanies(@RequestHeader ("Authorization") String token) {
+        if(authorizeClient.authorize(token)) {
+            List<CompanyResponseDTO> companies = companyService.getAllCompanies();
+            return ResponseEntity.ok(companies);
+        }
+        else
+        {
+            return new ResponseEntity<>("User not Authenticated",HttpStatus.BAD_REQUEST);
+        }
     }
 
     @PostMapping("/register")
-    public ResponseEntity<Company> RegisterCompany(@RequestBody Company company) {
-        Company c = companyService.RegisterCompany(company);
-        return ResponseEntity.ok(c);
+    public ResponseEntity<?> RegisterCompany(@RequestBody Company company,@RequestHeader ("Authorization") String token) {
+        if((authorizeClient.authorize(token)) && (authorizeClient.getrole(token).equalsIgnoreCase("admin")))
+        {
+            Company c = companyService.RegisterCompany(company);
+            return ResponseEntity.ok(c);
+        }
+        else
+        {
+        return new ResponseEntity<>("Sorry,Can only be Accessed by Admin",HttpStatus.BAD_REQUEST);
+        }
     }
 
     @PutMapping("/put/{companyCode}")
-    public ResponseEntity<String> updateCompany(@PathVariable long companyCode,@RequestBody Company company)
+    public ResponseEntity<String> updateCompany(@PathVariable long companyCode,@RequestBody Company company,@RequestHeader ("Authorization") String token)
     {
         try
         {
-            companyService.UpdateCompany(companyCode,company);
-            return ResponseEntity.ok("Company Details Updated Successfully");
+            if((authorizeClient.authorize(token)) && (authorizeClient.getrole(token).equalsIgnoreCase("admin"))) {
+                companyService.UpdateCompany(companyCode, company);
+                return ResponseEntity.ok("Company Details Updated Successfully");
+            }
+            else
+            {
+                return new ResponseEntity<>("Sorry,Can only be Accessed by Admin",HttpStatus.BAD_REQUEST);
+            }
         }
         catch (CompanyNotFound e)
         {
@@ -55,8 +84,14 @@ public class CompanyController {
     }
 
     @DeleteMapping("/delete/{companyCode}")
-    public ResponseEntity<Boolean> DeleteCompany(@PathVariable long companyCode)
+    public ResponseEntity<?> DeleteCompany(@PathVariable long companyCode,@RequestHeader ("Authorization") String token)
     {
-        return ResponseEntity.ok(companyService.deleteCompany(companyCode));
+        if((authorizeClient.authorize(token)) && (authorizeClient.getrole(token).equalsIgnoreCase("admin"))) {
+            return ResponseEntity.ok(companyService.deleteCompany(companyCode));
+        }
+        else
+        {
+            return new ResponseEntity<>("Sorry,Can only be Accessed by Admin",HttpStatus.BAD_REQUEST);
+        }
     }
 }
